Fixed a bug where the game could not be joined when restarting after creating an Arena Fixed particle enum bug when not selecting any trail in version 1.13 and higher

+++ src/main/java/com/lielamar/armsrace/listeners/shop/OnShopClick.java
@@ -80,7 +80,7 @@ public class OnShopClick implements Listener {
 						return;
 					}
 
-					double price = item.getPrice();
+					int price = item.getPrice();
 
 					if (main.getPlayerManager().getEconomy().withdrawPlayer(p, price).transactionSuccess()) {
 						cp.setTrails(item.getTrailType(), true);
@@ -110,7 +110,7 @@ public class OnShopClick implements Listener {
 						return;
 					}
 
-					double price = item.getPrice();
+					int price = item.getPrice();
 
 					if (main.getPlayerManager().getEconomy().withdrawPlayer(p, price).transactionSuccess()) {
 						cp.setKillEffects(item.getKillEffectType(), true);
